pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ntl-systoolbox"
version = "1.0.0"
description = "Outil CLI multiplateforme pour diagnostic, sauvegarde et audit d'obsolescence"
readme = "README.md"
authors = [
    {name = "NordTransit Logistics DSI", email = "dsi@nordtransit.local"}
]
license = {text = "Proprietary"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pyyaml>=6.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
ntl-systoolbox = "ntlsystoolbox.cli:main"

[project.urls]
Homepage = "https://github.com/nordtransit/ntl-systoolbox"
Documentation = "https://github.com/nordtransit/ntl-systoolbox/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ntlsystoolbox = ["config/*.yml"]
